[mod]: fixed watcher.module.ts
